+++ src/openrct2-ui/windows/Map.cpp
@@ -22,7 +22,6 @@
 #include <openrct2/OpenRCT2.h>
 #include <openrct2/actions/LandSetRightsAction.h>
 #include <openrct2/actions/MapChangeSizeAction.h>
-#include <openrct2/actions/ParkEntrancePlaceAction.h>
 #include <openrct2/actions/PeepSpawnPlaceAction.h>
 #include <openrct2/actions/SurfaceSetStyleAction.h>
 #include <openrct2/audio/audio.h>
@@ -38,7 +37,6 @@
 #include <openrct2/ride/Track.h>
 #include <openrct2/ride/TrainManager.h>
 #include <openrct2/ride/Vehicle.h>
-#include <openrct2/world/Entrance.h>
 #include <openrct2/world/Footpath.h>
 #include <openrct2/world/Scenery.h>
 #include <openrct2/world/Surface.h>
@@ -128,11 +126,8 @@ namespace OpenRCT2::Ui::Windows
         WIDX_CONSTRUCTION_RIGHTS_OWNED_CHECKBOX,
         WIDX_LAND_SALE_CHECKBOX,
         WIDX_CONSTRUCTION_RIGHTS_SALE_CHECKBOX,
-        WIDX_ROTATE_90,
         WIDX_MAP_GENERATOR
     };
-
-    validate_global_widx(WC_MAP, WIDX_ROTATE_90);
 
     // clang-format off
 static Widget window_map_widgets[] = {
@@ -233,8 +228,6 @@ static Widget window_map_widgets[] = {
         std::vector<uint8_t> _mapImageData;
         bool _mapWidthAndHeightLinked{ true };
         bool _recalculateScrollbars = false;
-        std::vector<Dropdown::Item> _entranceTypeDropdown{};
-        ObjectEntryIndex _selectedEntranceType = 0;
         enum class ResizeDirection
         {
             Both,
@@ -274,8 +267,6 @@ static Widget window_map_widgets[] = {
 
             // Reset land rights tool size
             _landRightsToolSize = 1;
-
-            InitParkEntranceDropdownItems();
         }
 
         void OnClose() override
@@ -340,10 +331,7 @@ static Widget window_map_widgets[] = {
                     Invalidate();
                     break;
                 case WIDX_BUILD_PARK_ENTRANCE:
-                    BuildParkEntranceToggle();
-                    break;
-                case WIDX_ROTATE_90:
-                    gWindowSceneryRotation = (gWindowSceneryRotation + 1) & 3;
+                    ContextOpenWindow(WindowClass::EditorOwnership);
                     break;
                 case WIDX_PEOPLE_STARTING_POSITION:
                     if (ToolSet(*this, widgetIndex, Tool::UpArrow))
@@ -413,9 +401,6 @@ static Widget window_map_widgets[] = {
 
                     Invalidate();
                     break;
-                case WIDX_BUILD_PARK_ENTRANCE:
-                    ShowParkEntranceDropdown(widgets[widgetIndex]);
-                    break;
             }
         }
 
@@ -475,9 +460,6 @@ static Widget window_map_widgets[] = {
                 case WIDX_SET_LAND_RIGHTS:
                     SetLandRightsToolUpdate(screenCoords);
                     break;
-                case WIDX_BUILD_PARK_ENTRANCE:
-                    PlaceParkEntranceToolUpdate(screenCoords);
-                    break;
                 case WIDX_PEOPLE_STARTING_POSITION:
                     SetPeepSpawnToolUpdate(screenCoords);
                     break;
@@ -488,9 +470,6 @@ static Widget window_map_widgets[] = {
         {
             switch (widgetIndex)
             {
-                case WIDX_BUILD_PARK_ENTRANCE:
-                    PlaceParkEntranceToolDown(screenCoords);
-                    break;
                 case WIDX_PEOPLE_STARTING_POSITION:
                     SetPeepSpawnToolDown(screenCoords);
                     break;
@@ -518,13 +497,6 @@ static Widget window_map_widgets[] = {
             switch (widgetIndex)
             {
                 case WIDX_SET_LAND_RIGHTS:
-                    Invalidate();
-                    HideGridlines();
-                    HideLandRights();
-                    HideConstructionRights();
-                    break;
-                case WIDX_BUILD_PARK_ENTRANCE:
-                    ParkEntranceRemoveGhost();
                     Invalidate();
                     HideGridlines();
                     HideLandRights();
@@ -567,99 +539,6 @@ static Widget window_map_widgets[] = {
             gMapSelectPositionB.x = mapCoords.x + size;
             gMapSelectPositionB.y = mapCoords.y + size;
             MapInvalidateSelectionRect();
-        }
-
-        CoordsXYZD PlaceParkEntranceGetMapPosition(const ScreenCoordsXY& screenCoords)
-        {
-            CoordsXYZD parkEntranceMapPosition{ 0, 0, 0, INVALID_DIRECTION };
-            const CoordsXY mapCoords = ViewportInteractionGetTileStartAtCursor(screenCoords);
-            parkEntranceMapPosition = { mapCoords.x, mapCoords.y, 0, INVALID_DIRECTION };
-            if (parkEntranceMapPosition.IsNull())
-                return parkEntranceMapPosition;
-
-            auto surfaceElement = MapGetSurfaceElementAt(mapCoords);
-            if (surfaceElement == nullptr)
-            {
-                parkEntranceMapPosition.SetNull();
-                return parkEntranceMapPosition;
-            }
-
-            parkEntranceMapPosition.z = surfaceElement->GetWaterHeight();
-            if (parkEntranceMapPosition.z == 0)
-            {
-                parkEntranceMapPosition.z = surfaceElement->GetBaseZ();
-                if ((surfaceElement->GetSlope() & kTileSlopeRaisedCornersMask) != 0)
-                {
-                    parkEntranceMapPosition.z += 16;
-                    if (surfaceElement->GetSlope() & kTileSlopeDiagonalFlag)
-                    {
-                        parkEntranceMapPosition.z += 16;
-                    }
-                }
-            }
-            parkEntranceMapPosition.direction = (gWindowSceneryRotation - GetCurrentRotation()) & 3;
-            return parkEntranceMapPosition;
-        }
-
-        void PlaceParkEntranceToolUpdate(const ScreenCoordsXY& screenCoords)
-        {
-            MapInvalidateSelectionRect();
-            MapInvalidateMapSelectionTiles();
-            gMapSelectFlags &= ~MAP_SELECT_FLAG_ENABLE;
-            gMapSelectFlags &= ~MAP_SELECT_FLAG_ENABLE_ARROW;
-            gMapSelectFlags &= ~MAP_SELECT_FLAG_ENABLE_CONSTRUCT;
-            CoordsXYZD parkEntrancePosition = PlaceParkEntranceGetMapPosition(screenCoords);
-            if (parkEntrancePosition.IsNull())
-            {
-                ParkEntranceRemoveGhost();
-                return;
-            }
-
-            int32_t sideDirection = (parkEntrancePosition.direction + 1) & 3;
-            gMapSelectionTiles.clear();
-            gMapSelectionTiles.push_back({ parkEntrancePosition.x, parkEntrancePosition.y });
-            gMapSelectionTiles.push_back({ parkEntrancePosition.x + CoordsDirectionDelta[sideDirection].x,
-                                           parkEntrancePosition.y + CoordsDirectionDelta[sideDirection].y });
-            gMapSelectionTiles.push_back({ parkEntrancePosition.x - CoordsDirectionDelta[sideDirection].x,
-                                           parkEntrancePosition.y - CoordsDirectionDelta[sideDirection].y });
-
-            gMapSelectArrowPosition = parkEntrancePosition;
-            gMapSelectArrowDirection = parkEntrancePosition.direction;
-
-            gMapSelectFlags |= MAP_SELECT_FLAG_ENABLE_CONSTRUCT | MAP_SELECT_FLAG_ENABLE_ARROW;
-            MapInvalidateMapSelectionTiles();
-            if (gParkEntranceGhostExists && parkEntrancePosition == gParkEntranceGhostPosition)
-            {
-                return;
-            }
-
-            ParkEntranceRemoveGhost();
-
-            auto gameAction = ParkEntrancePlaceAction(parkEntrancePosition, gFootpathSelectedId, _selectedEntranceType);
-            gameAction.SetFlags(GAME_COMMAND_FLAG_GHOST);
-
-            auto result = GameActions::Execute(&gameAction);
-            if (result.Error == GameActions::Status::Ok)
-            {
-                gParkEntranceGhostPosition = parkEntrancePosition;
-                gParkEntranceGhostExists = true;
-            }
-        }
-
-        void PlaceParkEntranceToolDown(const ScreenCoordsXY& screenCoords)
-        {
-            ParkEntranceRemoveGhost();
-
-            CoordsXYZD parkEntrancePosition = PlaceParkEntranceGetMapPosition(screenCoords);
-            if (!parkEntrancePosition.IsNull())
-            {
-                auto gameAction = ParkEntrancePlaceAction(parkEntrancePosition, gFootpathSelectedId, _selectedEntranceType);
-                auto result = GameActions::Execute(&gameAction);
-                if (result.Error == GameActions::Status::Ok)
-                {
-                    Audio::Play3D(Audio::SoundId::PlaceItem, result.Position);
-                }
-            }
         }
 
         void SetPeepSpawnToolUpdate(const ScreenCoordsXY& screenCoords)
@@ -908,8 +787,6 @@ static Widget window_map_widgets[] = {
             widgets[WIDX_SET_LAND_RIGHTS].bottom = height - 70 + 23;
             widgets[WIDX_BUILD_PARK_ENTRANCE].top = height - 46;
             widgets[WIDX_BUILD_PARK_ENTRANCE].bottom = height - 46 + 23;
-            widgets[WIDX_ROTATE_90].top = height - 46;
-            widgets[WIDX_ROTATE_90].bottom = height - 46 + 23;
             widgets[WIDX_PEOPLE_STARTING_POSITION].top = height - 46;
             widgets[WIDX_PEOPLE_STARTING_POSITION].bottom = height - 46 + 23;
 
@@ -941,13 +818,6 @@ static Widget window_map_widgets[] = {
 
             if ((gScreenFlags & SCREEN_FLAGS_SCENARIO_EDITOR) || GetGameState().Cheats.SandboxMode)
             {
-                // scenario editor: build park entrance selected, show rotate button
-                if ((InputTestFlag(INPUT_FLAG_TOOL_ACTIVE)) && gCurrentToolWidget.window_classification == WindowClass::Map
-                    && gCurrentToolWidget.widget_index == WIDX_BUILD_PARK_ENTRANCE)
-                {
-                    widgets[WIDX_ROTATE_90].type = WindowWidgetType::FlatBtn;
-                }
-
                 // Always show set land rights button
                 widgets[WIDX_SET_LAND_RIGHTS].type = WindowWidgetType::FlatBtn;
 
@@ -1051,22 +921,6 @@ static Widget window_map_widgets[] = {
         {
             InitMap();
             CentreMapOnViewPoint();
-        }
-
-        void OnDropdown(WidgetIndex widgetIndex, int32_t selectedIndex) override
-        {
-            if (selectedIndex == -1)
-            {
-                return;
-            }
-
-            switch (widgetIndex)
-            {
-                case WIDX_BUILD_PARK_ENTRANCE:
-                    BuildParkEntranceToggle();
-                    BuildParkEntranceClick(selectedIndex);
-                    break;
-            }
         }
 
     private:
@@ -1598,57 +1452,6 @@ static Widget window_map_widgets[] = {
             }
             width = std::max(min_width, width);
             _recalculateScrollbars = true;
-        }
-
-        void InitParkEntranceDropdownItems()
-        {
-            _entranceTypeDropdown.clear();
-            for (auto objectIndex = 0; objectIndex < OBJECT_ENTRY_INDEX_NULL; objectIndex++)
-            {
-                auto& objManager = GetContext()->GetObjectManager();
-                auto* object = static_cast<EntranceObject*>(objManager.GetLoadedObject(ObjectType::ParkEntrance, objectIndex));
-                if (object != nullptr)
-                {
-                    const auto* legacyData = reinterpret_cast<const EntranceEntry*>(object->GetLegacyData());
-                    _entranceTypeDropdown.push_back({ .Format = legacyData->string_idx, .Args = objectIndex, .Flags = 0 });
-                }
-            }
-        }
-
-        void ShowParkEntranceDropdown(Widget& widget)
-        {
-            gDropdownDefaultIndex = 0;
-            auto numItems = _entranceTypeDropdown.size();
-            for (auto dropdownIndex = 0u; dropdownIndex < numItems; dropdownIndex++)
-            {
-                gDropdownItems[dropdownIndex] = _entranceTypeDropdown[dropdownIndex];
-
-                if (gDropdownItems[dropdownIndex].Args == _selectedEntranceType)
-                    gDropdownDefaultIndex = dropdownIndex;
-            }
-
-            WindowDropdownShowText(
-                { windowPos.x + widget.left, windowPos.y + widget.top }, widget.height() + 1,
-                ColourWithFlags(colours[1]).withFlag(ColourFlag::translucent, true), 0, numItems);
-        }
-
-        void BuildParkEntranceToggle()
-        {
-            Invalidate();
-            if (ToolSet(*this, WIDX_BUILD_PARK_ENTRANCE, Tool::UpArrow))
-                return;
-
-            gParkEntranceGhostExists = false;
-            InputSetFlag(INPUT_FLAG_6, true);
-
-            ShowGridlines();
-            ShowLandRights();
-            ShowConstructionRights();
-        }
-
-        void BuildParkEntranceClick(int16_t dropdownIndex)
-        {
-            _selectedEntranceType = gDropdownItems[dropdownIndex].Args;
         }
     };
