+++ src/openrct2/interface/Window.h
@@ -358,7 +358,7 @@ constexpr int32_t WC_PEEP__WIDX_ACTION_LBL = 13;
 constexpr int32_t WC_PEEP__WIDX_PICKUP = 14;
 constexpr int32_t WC_TRACK_DESIGN_LIST__WIDX_ROTATE = 8;
 constexpr int32_t WC_TRACK_DESIGN_PLACE__WIDX_ROTATE = 3;
-constexpr int32_t WC_MAP__WIDX_ROTATE_90 = 24;
+constexpr int32_t WC_EDITOR_OWNERSHIP__WIDX_ROTATE_ENTRANCE_BUTTON = 6;
 constexpr int32_t WC_EDITOR_OBJECT_SELECTION__WIDX_TAB_1 = 22;
 constexpr int32_t WC_STAFF__WIDX_PICKUP = 9;
 constexpr int32_t WC_TILE_INSPECTOR__WIDX_BUTTON_ROTATE = 13;
